+++ RestaurantApp/app/src/main/java/com/example/restaurantapp/fragments/ListFragment.kt
@@ -241,7 +241,7 @@ class ListFragment : Fragment(), RestaurantAdapter.OnItemClickListener {
 
         val bundle = Bundle()
 
-        bundle.putInt("id", restaurant.id)
+        bundle.putLong("id", restaurant.id)
         details.arguments = bundle
 
         activity!!.supportFragmentManager.beginTransaction()
